mlp_api/requirements.txt
Django==2.2
django-tinymce==2.8.0
djangorestframework==3.9.2
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
python-slugify==3.0.2
pytils==0.3
pytz==2019.1
sqlparse==0.3.0
text-unidecode==1.2
